sdk: xmldoc pass on the resolver/static-semantics work + split directive nodes
Two things flagged in review of #182-184:
- One-liner <summary>text</summary> tags, and a DetermineDeclaredType with
  <param>/<exception> but no <summary> at all. Fixed across the #174/#175
  files (LexicalScope, ScopeTree, LexicalScopeKind, SymbolResolutionResult,
  ModuleDirectives, VBModuleSymbol, ITypedSymbol, SimpleNameExpressionStaticSemantics)
  to match the established multi-line style.
- RDCore.SDK/Model/AST/Directives/ModuleDirectives.cs grouped four unrelated
  directive node types (AttributeDirectiveNode, ModuleOptionDirectiveNode,
  TypeDefDirectiveNode, ImplementsDirectiveNode) under a filename matching
  none of them — surprising, and now doubly so next to the real
  ModuleDirectives type in Model/Symbols/. Split into one file per type.
